Reorganized project Added maven
