+++ services/src/main/java/org/services/StudentServiceImpl.java
@@ -1,4 +1,7 @@
-package ru.vsu.implementation;
+package org.services;
+
+import org.practice.ExamResult;
+import org.practice.Student;
 
 import java.util.Collection;
 import java.util.Comparator;
@@ -6,11 +9,8 @@ import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import ru.vsu.entity.ExamResult;
-import ru.vsu.entity.Student;
-
-public class StudentService implements ru.vsu.logic.StudentService {
-    public StudentService() {
+public class StudentServiceImpl implements org.services.StudentService {
+    public StudentServiceImpl() {
     }
 
     @Override
